requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-Cors==4.0.0
requests==2.31.0
python-multipart==0.0.6
Pillow>=9.0.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-translation-text==1.0.0b1
Flask-Session==0.5.0
Werkzeug==2.3.7
gunicorn==21.2.0
